% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Autogen.R
\name{Autogen}
\alias{Autogen}
\title{Autogen will be called from ConvertandgraphIns() and be used to automatically create 4 heatmaps after conversion.}
\usage{
Autogen(dataset, mapname, string)
}
\arguments{
\item{dataset}{The data set to graph.}

\item{mapname}{The name of the raster image to plot against.}

\item{string}{The text you wish to display on the image.}
}
\description{
This function will automatically return a plot from the given data. Must use converted data. Set for 1024 Insurgency image defaults.
}

