% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertandgraphIns.R
\name{ConvertandgraphIns}
\alias{ConvertandgraphIns}
\title{Convert and Graph Insrugency Data}
\usage{
ConvertandgraphIns(dataset, mapname, env = globalenv(), Display = TRUE)
}
\arguments{
\item{dataset}{The data set you wish to graph}

\item{mapname}{The map name you wish to graph the data against. This is used to get the conversion for the data set. (conversion is map specific)}

\item{env}{Set the enviroment in wish to save the converted data sets.}

\item{Display}{will display each plot with a promt to continue. Set False to only save PDF of generated images.}
}
\value{
Will return multiple data sets to the global enviroment, as well as a PDF of all saved plots.
}
\description{
Will take in a data set, run a conversion based on the map name given, then return a series of datasets to then be used for graphing.
}

