% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateHeatMap.R
\name{CreateHeatMap}
\alias{CreateHeatMap}
\title{Plot For HeatMap Generation Allows full control over many variables.}
\usage{
CreateHeatMap(dataset, X = dataset[, 2], Y = dataset[, 3], xpixel = 1024,
  ypixel = 1024, Map, string, alphalev = 1)
}
\arguments{
\item{dataset}{The data set being used for graphic}

\item{X}{The column X is being stored}

\item{Y}{The column Y is being stored}

\item{xpixel}{The size of x background image in pixels that is being used. Defaults to 1024.}

\item{ypixel}{The size of y background image in pixels that is being used. Defaults to 1024.}

\item{Map}{gives the map name which is used to grab the correct image file from the enviroment. Use case sensative name of raster image. IE if image name is imarket type in "imarket" and not "Imarket" or "iMarket"}

\item{string}{is the title you wish to see at the top left of the image.}

\item{alphalev}{this is the value you wish to set the color alpha levels.}
}
\value{
The return object is the plot created by ggplot.
}
\description{
Function that uses ggplot to create the basic geom_point with given demensions (default 1024 for Insurgency) with theme_nothing() from cowplot.
}

