% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateMapRaster.R
\name{CreateMapRaster}
\alias{CreateMapRaster}
\title{CreateMapRaster}
\usage{
CreateMapRaster(Image = file.choose(), string, env = globalenv())
}
\arguments{
\item{Image}{Set address of Image file, file.choose() by default.}

\item{string}{Set name you wish the image to be called.}

\item{env}{Enviroment set to global so user will see it.}
}
\description{
Will take in an image location (file.choose() default), String to call the object, and Enviroment location (Global defaul). This allows the user to create a custom raster image
}

