% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomconvertIns.R
\name{CustomconvertIns}
\alias{CustomconvertIns}
\title{CustomconvertIns}
\usage{
CustomconvertIns(dataset, mapname, Posx, Posy, scale, env = globalenv())
}
\arguments{
\item{dataset}{The data you wish to convert}

\item{mapname}{Name of the map}

\item{Posx}{Posx offset}

\item{Posy}{Posy offset}

\item{scale}{scale value}

\item{env}{set the enviroment, Global by default.}
}
\value{
Does not return a value does create multiple data sets.
}
\description{
Forumla to take in a raw data set and convert it from given values.
}

