% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iheights.R
\docType{data}
\name{iheights}
\alias{iheights}
\title{RasterGrob image of buhriz. Used as background image for Heatmap.}
\format{An object of class \code{rastergrob} (inherits from \code{grob}, \code{gDesc}) of length 12.}
\source{
unpacked from game files of INS under Overview of parsed map information.
}
\usage{
iheights
}
\description{
Image is called ibuhriz.
}
\keyword{datasets}

