% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iministry.R
\docType{data}
\name{iministry}
\alias{iministry}
\title{RasterGrob image of Ministry. Used as background image for Heatmap.}
\format{An object of class \code{rastergrob} (inherits from \code{grob}, \code{gDesc}) of length 12.}
\source{
unpacked from game files of INS under Overview of parsed map information.
}
\usage{
iministry
}
\description{
Image is called iministry.
}
\keyword{datasets}

