% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rawdata.R
\docType{data}
\name{Rawdata}
\alias{Rawdata}
\title{Example Rawdata set that contains data for both kills and deaths for Market.}
\format{An object of class \code{data.frame} with 5469 rows and 10 columns.}
\source{
unpacked from game files of INS under Overview of parsed map information.
}
\usage{
Rawdata
}
\description{
Example Rawdata set that contains data for both kills and deaths for Market.
}
\keyword{datasets}

